+++ lib/config/schema.js
@@ -3,7 +3,7 @@
 /**
  * Configuration options schema.
  * This is used by `ajv-cli` to compile the validation function
- * used internally by Doc.
+ * used internally by `doc`.
  * The `validate.js` file containing the function is generated with the
  * `npm run build` script.
  */
