+++ lib/eventLoopDelay.js
@@ -17,7 +17,7 @@ class EventLoopDelayMetric {
       this[kRawMetric] = monitorEventLoopDelay(this[kOptions])
       this[kRawMetric].enable()
     } else {
-      this[kRawMetric] = 0.0
+      this[kRawMetric] = 0
     }
   }
 
@@ -30,9 +30,9 @@ class EventLoopDelayMetric {
   }
 
   compute (raw) {
-    // In some cases the mean value is NaN, so I am forcing it
-    // to be a float to keep V8 from deoptimizing this path.
-    const value = isNaN(raw) ? 0.0 : raw
+    // In some cases the mean value is NaN because
+    // there are not enough samples.
+    const value = isNaN(raw) ? 0 : raw
     const delta = value / 1e6 - this[kOptions].resolution
     return Math.max(0, delta)
   }
