+++ package.json
@@ -13,8 +13,10 @@
     "test": "npm run lint && npm run test:js && npm run test:ts",
     "test:js": "tap",
     "test:ts": "tsd",
-    "deoptigate:server": "PORT=3000 deoptigate benchmarks/server.js",
+    "deoptigate:server": "PORT=3000 deoptigate benchmarks/doc.js",
     "deoptigate:load": "autocannon -d 60 -c 100 -p 10 localhost:3000",
+    "bench:base": "PORT=3000 concurrently -k -s first \"node benchmarks/base.js\" \"node -e 'setTimeout(()=>{}, 1000)' && autocannon -d 60 -c 100 -p 10 localhost:3000\"",
+    "bench:doc": "PORT=3000 concurrently -k -s first \"node benchmarks/doc.js\" \"node -e 'setTimeout(()=>{}, 1000)' && autocannon -d 60 -c 100 -p 10 localhost:3000\"",
     "doc": "markdown-toc -i README.md",
     "prerelease": "npm cit",
     "release": "HUSKY_SKIP_HOOKS=1 standard-version",
